separate Handles and Local types into their own files
